requirements.txt
yfinance
pandas
fredapi
tensorflow
numpy
python-dotenv
keras-ncp
seaborn
scikit-learn
matplotlib
statsmodels
ipykernel